server/openapi_server/test/test_default_controller.py
import unittest

from flask import json

from openapi_server.models.app import App  # noqa: E501
from openapi_server.models.error import Error  # noqa: E501
from openapi_server.models.new_app import NewApp  # noqa: E501
from openapi_server.test import BaseTestCase


class TestDefaultController(BaseTestCase):
    """DefaultController integration test stubs"""

    def test_add_app(self):
        """Test case for add_app

        
        """
        new_app = {"name":"name","tag":"tag"}
        headers = { 
            'Accept': 'application/json',
            'Content-Type': 'application/json',
        }
        response = self.client.open(
            '/v1/Apps',
            method='POST',
            headers=headers,
            data=json.dumps(new_app),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_delete_app(self):
        """Test case for delete_app

        
        """
        headers = { 
            'Accept': 'application/json',
        }
        response = self.client.open(
            '/v1/Apps/{id}'.format(id=56),
            method='DELETE',
            headers=headers)
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_find_app_by_id(self):
        """Test case for find_app_by_id

        
        """
        headers = { 
            'Accept': 'application/json',
        }
        response = self.client.open(
            '/v1/Apps/{id}'.format(id=56),
            method='GET',
            headers=headers)
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_find_apps(self):
        """Test case for find_apps

        
        """
        query_string = [('tags', ['tags_example']),
                        ('limit', 56)]
        headers = { 
            'Accept': 'application/json',
        }
        response = self.client.open(
            '/v1/Apps',
            method='GET',
            headers=headers,
            query_string=query_string)
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))


if __name__ == '__main__':
    unittest.main()
